# Update perf report to show the number of operators and per operator avg time in summary